% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{data}
\name{AJR}
\alias{AJR}
\title{AJR data set}
\format{
\describe{ 
\item{Mort}{Settler mortality}
\item{logMort}{logarithm of Mort}
\item{Latitude}{Latitude}
\item{Latitude2}{Latitude^2}
\item{Africa}{Africa}
\item{Asia}{Asia}
\item{Namer}{North America}
\item{Samer}{South America}
\item{Neo}{Neo-Europes}
\item{GDP}{GDP}
\item{Exprop}{Average protection against expropriation risk} 
}
}
\description{
Dataset on settler mortality.
}
\details{
Data set was analysed in Acemoglu et al. (2001). A detailed description of the data can be found at \url{https://economics.mit.edu/people/faculty/daron-acemoglu/data-archive}
}
\examples{
data(AJR)
}
\references{
D. Acemoglu, S. Johnson, J. A. Robinson  (2001). Colonial origins of comparative development: an empirical investigation.
American Economic Review, 91, 1369--1401.
}
\keyword{datasets}
