% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialROC.R
\name{AreaPredictedPresence}
\alias{AreaPredictedPresence}
\title{Helper function to compute the area (number of pixels) that a certain threshold has.}
\usage{
AreaPredictedPresence(InRast)
}
\arguments{
\item{InRast}{A raster class object of a continuous model output.}
}
\description{
Helper function to compute the area (number of pixels) that a certain threshold has.
}
