% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialROC.R
\name{auc_comp}
\alias{auc_comp}
\title{Helper function to compute partial AUC, AUC ratio.}
\usage{
auc_comp(IterationNo, OccurTbl, RandomPercent, OmissionVal, ClassPixels)
}
\arguments{
\item{IterationNo}{number of iteration to compute partial AUC values}

\item{OccurTbl}{Validation data. Must have 3 columns SpName, Longitude, Latitude.}

\item{RandomPercent}{Occurrence points to be sampled randomly from the test data for bootstrapping.}

\item{OmissionVal}{1-E.}

\item{ClassPixels}{Pixel classes.}
}
\description{
Helper function to compute partial AUC, AUC ratio.
}
