% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{data}
\name{BLP}
\alias{BLP}
\title{BLP data set}
\format{
\describe{ 
\item{model.name}{model name}
\item{model.id}{model id}
\item{firm.id}{firm id}
\item{cdid}{cdid}
\item{id}{id}
\item{price}{log price}
\item{mpg}{miles per gallon}
\item{mpd}{miles per dollar}
\item{hpwt}{horse power per weight}
\item{air}{air conditioning (binary variable)}
\item{space}{size of the car}
\item{share}{market share}
\item{outshr}{share s0}
\item{y}{outcome variable defined as log(share) - log(outshr) }
\item{trend}{time trend}
}
}
\description{
Automobile data set from the US.
}
\details{
Data set was analysed in Berry, Levinsohn and Pakes (1995). The data stem from annual issues of the Automotive News Market Data Book. 
The data set inlcudes information on all models marketed during the the period beginning 1971 and ending in 1990 cotaining 2217 model/years from 997 distinct models.
A detailed description is given in BLP (1995, 868--871). The internal function \code{constructIV} constructs instrumental variables along the lines described and used in BLP (1995).
}
\examples{
data(BLP)
}
\references{
S. Berry, J. Levinsohn, A. Pakes (1995). Automobile Prices in Market EquilibriumD. Econometrica, 63(4), 841--890.
}
\keyword{datasets}
