% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialROC.R
\name{CalculateAUC}
\alias{CalculateAUC}
\title{Helper function to compute AUC (partialAUC, AUC at Random, AUC ratio) values}
\usage{
CalculateAUC(XYTable, OmissionVal, IterationNo)
}
\arguments{
\item{XYTable}{A table with the output of the function GenerateXYTableb}

\item{OmissionVal}{Omission value.}

\item{IterationNo}{Number of boostrap interation.}
}
\description{
Helper function to compute AUC (partialAUC, AUC at Random, AUC ratio) values
}
