% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dup_by_year.R
\name{clean_dup_by_year}
\alias{clean_dup_by_year}
\title{Function to clean duplicated coordinates data}
\usage{
clean_dup_by_year(this.species, threshold)
}
\arguments{
\item{this.species, }{Species Temporal Data object see \code{\link[hdm]{sp_temporal_data}}.}

\item{threshold}{A numerc value representig the distance (in grades) between coordinates
to be considered as a duplicate.}
}
\description{
Clean duplicated longitude and latitude data by year using threshold distance
             which is a distance
}
