% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{data}
\name{cps2012}
\alias{cps2012}
\title{cps2012 data set}
\format{\describe{ 
\item{lnw}{log of hourly wage (annual earnings / annual hours)}
\item{female}{female indicator}
\item{married status}{ six indicators: widowed, divorced, separated, nevermarried, and married (omitted)}
\item{education attainment}{six indicators: hsd08, hsd911, hsg, cg, ad, and sc (omitted)}
\item{region indicators}{four indicators: mw, so, we, and ne (omitted)}
\item{potential experience}{(max[0, age - years of education - 7]): exp1, exp2 (divided by 100), exp3 (divided by 1000), exp4 (divided by 10000)}
\item{weight}{March Supplement sampling weight}
\item{year}{CPS year}
}}
\description{
Census data from the US for the year 2012.
}
\details{
The CPS is a monthly U.S. household survey conducted jointly by the U.S. Census Bureau and the Bureau of Labor Statistics. The data comprise the year 2012. 
This data set was used in Mulligan and Rubinstein (2008). 
The sample comprises white non-hipanic, ages 25-54, working full time full year (35+ hours per week at least 50 weeks), exclude living in group quarters, 
 self-employed, military, agricultural, and private household sector, allocated earning, inconsistent report on earnings and employment, missing data.
}
\examples{
data(BLP)
}
\references{
C. B. Mulligan and Y. Rubinstein (2008). Selection, investment, and women's relative wages over time. The Quarterly Journal of Economics, 1061--1110.
}
\keyword{datasets}
