% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoDebiased_backends.R
\name{DataAPDAutoDML}
\alias{DataAPDAutoDML}
\title{Backend for rlassoAPDAutoDML}
\usage{
DataAPDAutoDML(
  x,
  d,
  y,
  x_manual = NULL,
  data,
  poly_order = 3,
  interactions = TRUE,
  unit = NULL,
  time = NULL
)
}
\arguments{
\item{x}{\code{character} specifying exogenous variables}

\item{d}{\code{character} specifying treatment variable (continuous)}

\item{y}{\code{character} specifying outcome variable / dependent variable.}

\item{x_manual}{\code{character} specifying variables that should manually be added to the model. These variables are excluded from the the internal preprocessing w.r.t. to polynomials and interactions. The corresponding data will be adjusted during derivative calculations. Variables should not include interactions with variable \code{d}.}

\item{data}{\code{data.frame} or \code{data.table} with data set}

\item{poly_order}{Polynomial order.}

\item{interactions}{\code{logical} specifying whether two-way interactions of \code{x} and \code{d} should be constructed. Default is \code{true}.}

\item{unit}{\code{character} specifying unit variable of data in panel context, if not panel data omit (default \code{NULL}).}

\item{time}{\code{character} specifying time variable of data in panel context, if not panel data omit (default \code{NULL}).}
}
\value{
an object of class \code{DataAPDAutoDML} with the following entries
\itemize{
  \item{\code{data_prep} - \code{data.table} object containing the preprocessed data}
  \item{\code{Y} - Y as passed as input} 
  \item{\code{X} - Processed X (dictionary applied to input data)}
  \item{\code{p} - Number of columns of preprocessed matrix \code{X}}
  \item{\code{M} - Matrix M as of ... TODO }
}
}
\description{
Data backend for auto-debiased estimation of average partial derivatives 
(continuous treatment variable). Implements the preprocessing steps required 
for auto-debiased estimation of the APD. If interested in using panel data, 
user must also enter "unit" and "time" variables as well. If using cross section
these fields can be left empty (\code{NULL}).
}
