% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoDebiased_backends.R
\name{DataATEAutoDML}
\alias{DataATEAutoDML}
\title{Backend for rlassoATEAutoDML}
\usage{
DataATEAutoDML(x, d, y, data, dict = NULL)
}
\arguments{
\item{x}{\code{character} specifying exogenous variables}

\item{d}{\code{character} specifying treatment variable (binary)}

\item{y}{\code{character} specifying outcome variable / dependent variable}

\item{data}{\code{data.frame} or \code{data.table} containing the raw input data.}

\item{dict}{a dictionary (\code{function})
function of (D,X) with D being a numeric and X a matrix, which returns a numeric vector. Default is \code{dict <- function(d, x) return(c(1, d, x))}. The dictionary will be applied on an individual level, i.e., being executed for each i in 1, ..., n with i representing a row in the \code{data}.}
}
\value{
an object of class \code{DataAutoDebiasedATE} with the following entries
\itemize{
  \item{\code{data_prep} - A copy of the original \code{data} (containing only variables as provided through \code{x}, \code{d} and \code{y})}
  \item{\code{Y} - Y as passed as input} 
  \item{\code{X} - Processed X (dictionary applied to input data)}
  \item{\code{p} - Number of columns of preprocessed matrix \code{X}}
  \item{\code{M} - Matrix M as of ... TODO }
  \item{\code{X_prelim} - Version of matrix X used for a preliminary estimator}
  \item{\code{M_prelim} - Version of matrix M used for a preliminary estimator}
  \item{\code{p_prelim_out} - Number of columns of preprocessed matrix \code{X_prelim}}
}
}
\description{
Data backend for auto debiased ATE. Implements the preprocessing steps required for estimation of the ATE using auto debiased ATE.
}
