% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_area.R
\name{distribution_area}
\alias{distribution_area}
\title{Function to compute the total area of distribution of a binary raster.}
\usage{
distribution_area(binaryMap)
}
\arguments{
\item{binaryMap}{A binary raster projected in decimal coordinates (longitude a and latitude).}
}
\description{
Function to compute the total area of distribution of a binary raster.
}
