% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_by_year.R
\name{extract_by_year}
\alias{extract_by_year}
\title{Extract environmental data by year.}
\usage{
extract_by_year(this.species, layers_pattern)
}
\arguments{
\item{this.species, }{Species Temporal Data object see \code{\link[hdm]{sp_temporal_data}}.}

\item{layers_pattern}{A regular expression to indicate the pattern to search in layers files names.}
}
\description{
Extract environmental data by year.
}
