% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML_helpers.R
\name{for_tex}
\alias{for_tex}
\title{prints output of rlassoAutoDML in latex table format}
\usage{
for_tex(spec1)
}
\arguments{
\item{spec1}{output of rlassoAutoDML}
}
\description{
prints output of rlassoAutoDML in latex table format
}
