% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialROC.R
\name{GenerateXYTableb}
\alias{GenerateXYTableb}
\title{Helper function to compute the area (number of pixels) that a certain threshold has.}
\usage{
GenerateXYTableb(ClassPixels, OccurINClass)
}
\arguments{
\item{ClassPixels}{Pixel threshold class .}

\item{OccurINClass}{Ocurrence points that lies in a certain class.}
}
\description{
Helper function to compute the area (number of pixels) that a certain threshold has.
}
