% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{data}
\name{Growth Data}
\alias{Growth Data}
\alias{Growth}
\alias{Example}
\alias{GrowthData}
\alias{GDP}
\title{Growth data set}
\format{
Dataframe with the following variables:
\describe{
\item{outcome}{dependent variable: national growth rates in GDP per capita
for the periods 1965-1975 and 1975-1985}
\item{x}{covariates which might influence growth}}
}
\source{
The full data set and further details can be found at
\url{http://www.nber.org/pub/barro.lee}, \url{http://www.barrolee.com}, and,
\url{http://www.bristol.ac.uk//Depts//Economics//Growth//barlee.htm}.
}
\description{
Data set of growth compiled by Barro Lee.
}
\details{
The data set contains growth data of Barro-Lee. The Barro Lee data consists
of a panel of 138 countries for the period 1960 to 1985. The dependent
variable is national growth rates in GDP per capita for the periods
1965-1975 and 1975-1985. The growth rate in GDP over a period from \eqn{t_1} to \eqn{t_2}
is commonly defined as \eqn{\log(GDP_{t_1}/GDP_{t_2})}. The number of covariates is p=62.
The number of complete observations is 90.
}
\examples{
data(GrwothData)
}
\references{
R.J. Barro, J.W. Lee (1994). Data set for a panel of 139
countries. NBER.

R.J. Barro, X. Sala-i-Martin (1995). Economic Growth. McGrwa-Hill, New York.
}
\keyword{GDP}
\keyword{Grwoth}
\keyword{datasets}
