% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{package}
\name{hdm-package}
\alias{hdm-package}
\alias{hdm}
\title{hdm: High-Dimensional Metrics}
\description{
This package implements methods for estimation and inference in a high-dimensional setting.
}
\details{
\tabular{ll}{ Package: \tab hdm\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2015-05-25\cr License: \tab GPL-3\cr } This package provides efficient estimators 
and  uniformly valid confidence intervals for various low-dimensional causal/structural parameters  
appearing in high-dimensional approximately sparse models. The package
includes functions for fitting heteroskedastic robust Lasso regressions with non-Gaussian erros and 
for instrumental variable (IV) and treatment effect estimation in a
high-dimensional setting. Moreover, the methods enable valid post-selection
inference. Moreover, a theoretically grounded, data-driven choice of the penalty level is provided.
}
\references{
A. Belloni, D. Chen, V. Chernozhukov and C. Hansen (2012).
Sparse models and methods for optimal instruments with an application to
eminent domain. \emph{Econometrica} 80 (6), 2369-2429.

A. Belloni, V. Chernozhukov and C. Hansen (2013). Inference for
high-dimensional sparse econometric models. In Advances in Economics and
Econometrics: 10th World Congress, Vol. 3: Econometrics, Cambirdge
University Press: Cambridge, 245-295.

A. Belloni, V. Chernozhukov, C. Hansen (2014). Inference on
treatment effects after selection among high-dimensional controls. The
Review of Economic Studies 81(2), 608-650.
}
\author{
Victor Chernozhukov, Christian Hansen, Martin Spindler\cr

Maintainer: Martin Spindler <spindler@mea.mpisoc.mpg.de>
}
