% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{package}
\name{hdm2-package}
\alias{hdm2}
\alias{hdm2-package}
\title{hdm2: High-Dimensional Metrics}
\description{
This packages implements high-dimensional methods with a focus on
Econometrics.
}
\details{
\tabular{ll}{ Package: \tab hdm\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2015-05-25\cr License: \tab GPL-3\cr } The package
includes functions for fitting a heteroskedastic robust Lasso regression and
for instrumental variable (IV) and treatment effect estimation in a
high-dimensional setting. Moreover, the methods enable valid post-selection
inference.
}
\author{
Victor Chernozhukov, Christian Hansen, Martin Spindler\cr

Maintainer: Martin Spindler <spindler@mea.mpisoc.mpg.de>
}
\references{
TBD
}
\keyword{Endogeneity}
\keyword{Evaluation}
\keyword{Instrumental}
\keyword{Lasso}
\keyword{Microeconometrics}
\keyword{Program}
\keyword{Variables}
\keyword{effects}
\keyword{package}
\keyword{treatment}

