% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{print.rlasso}
\alias{print.rlasso}
\alias{methods.rlasso}
\alias{predict.rlasso}
\alias{model.matrix.rlasso}
\alias{summary.rlasso}
\alias{model.matrix.rlasso}
\alias{predict.rlasso}
\title{Methods for S3 object \code{rlasso}}
\usage{
\method{print}{rlasso}(x, all = TRUE, digits = max(3L, getOption("digits") -
  3L), ...)

\method{summary}{rlasso}(object, all = TRUE, digits = max(3L,
  getOption("digits") - 3L), ...)

\method{model.matrix}{rlasso}(object, ...)

\method{predict}{rlasso}(object, newdata = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{rlasso}}

\item{all}{logical, indicates if coefficients of all variables (TRUE) should be displayed or only the non-zero ones (FALSE)}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{an object of class \code{rlasso}}

\item{newdata}{new data set for prediction. An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are returned.}
}
\description{
Objects of class \code{rlasso} are constructed by \code{rlasso}.
\code{print.rlasso} prints and displays some information about fitted \code{rlasso} objects.
\code{summary.rlasso} summarizes information of a fitted \code{rlasso} object.
\code{predict.rlasso} predicts values based on a \code{rlasso} object.
\code{model.matrix.rlasso} constructs the model matrix of a \code{rlasso} object.
}
