% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoATEAutoDML.R
\name{print.rlassoATEAutoDML}
\alias{print.rlassoATEAutoDML}
\alias{methods.rlassoATEAutoDML}
\alias{summary.rlassoATEAutoDML}
\alias{confint.rlassoATEAutoDML}
\title{Methods for S3 object \code{rlassoATEAutoDML}}
\usage{
\method{print}{rlassoATEAutoDML}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoATEAutoDML}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{confint}{rlassoATEAutoDML}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class \code{rlassoATEAutoDML}}

\item{digits}{number of significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required.}
}
\description{
Objects of class \code{rlassoATEAutoDML} are constructed by  \code{rlassoATEAutoDML}.
\code{print.rlassoATEAutoDML} prints and displays some information about fitted \code{rlassoATEAutoDML} objects.
\code{summary.rlassoATEAutoDML} summarizes information of a fitted \code{rlassoATEAutoDML} object.
\code{confint.rlassoATEAutoDML} extracts the confidence intervals.
}
