% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML.R
\name{print.rlassoAutoDML}
\alias{print.rlassoAutoDML}
\alias{methods.rlassoAutoDML}
\alias{summary.rlassoAutoDML}
\alias{confint.rlassoAutoDML}
\title{Methods for S3 object \code{rlassoAutoDML}}
\usage{
\method{print}{rlassoAutoDML}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoAutoDML}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{confint}{rlassoAutoDML}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class \code{rlassoAutoDML}}

\item{digits}{number of significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required.}
}
\description{
Objects of class \code{rlassoAutoDML} are constructed by  \code{rlassoAutoDML}.
\code{print.rlassoAutoDML} prints and displays some information about fitted \code{rlassoAutoDML} objects.
\code{summary.rlassoAutoDML} summarizes information of a fitted \code{rlassoAutoDML} object.
\code{confint.rlassoAutoDML} extracts the confidence intervals.
}
