% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlassoEffect.R
\name{print.rlassoEffect}
\alias{confint.rlassoEffect}
\alias{methods.rlassoEffect}
\alias{plot.rlassoEffect}
\alias{print.rlassoEffect}
\alias{summary.rlassoEffect}
\title{Methods for S3 object \code{rlassoEffect}}
\usage{
\method{print}{rlassoEffect}(x, digits = max(3L, getOption("digits") - 3L),
  ...)

\method{summary}{rlassoEffect}(object, digits = max(3L, getOption("digits") -
  3L), ...)

\method{confint}{rlassoEffect}(object, parm, level = 0.95, joint = FALSE,
  ...)

\method{plot}{rlassoEffect}(x, main = "", xlab = "coef", ylab = "",
  xlim = NULL, col = "black", ...)
}
\arguments{
\item{x}{An object of class \code{rlassoEffect}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods.}

\item{object}{An object of class \code{rlassoEffect}}

\item{parm}{a specification of which parameters are to be given confidence intervals among the variables for which inference was done, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required}

\item{joint}{logical, if \code{TRUE} joint confidence intervals are clalculated.}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{xlim}{vector of length two giving lower and upper bound of x axis}

\item{col}{color of lines of the graph}
}
\description{
Objects of class \code{rlassoEffect} are constructed by \code{rlassoEffect.formula} or \code{rlassoEffect.default}.
\code{print.rlassoEffect} prints and displays some information about fitted \code{rlassoEffect} objects.
\code{summary.rlassoEffect} summarizes information of a fitted \code{rlassoEffect} object.
\code{confint.rlassoEffect} extracts the confidence intervals.
\code{plot.rlassoEffect} plots the estimates with confidence intervals.
}
\keyword{methods}
\keyword{rlassoEffect}

