% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoEffects.R
\name{print.rlassoEffects}
\alias{print.rlassoEffects}
\alias{methods.rlassoEffects}
\alias{confint.rlassoEffects}
\alias{plot.rlassoEffects}
\title{Methods for S3 object \code{rlassoEffects}}
\usage{
\method{print}{rlassoEffects}(x, digits = max(3L, getOption("digits") -
  3L), ...)

\method{confint}{rlassoEffects}(object, parm, level = 0.95,
  joint = FALSE, ...)

\method{plot}{rlassoEffects}(x, joint = FALSE, level = 0.95,
  main = "", xlab = "coef", ylab = "", xlim = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{rlassoEffects}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods.}

\item{object}{an object of class \code{rlassoEffects}}

\item{parm}{a specification of which parameters are to be given confidence intervals among the variables for which inference was done, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required}

\item{joint}{logical, if \code{TRUE} joint confidence intervals are calculated.}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{xlim}{vector of length two giving lower and upper bound of x axis}
}
\description{
Objects of class \code{rlassoEffects} are constructed by  \code{rlassoEffects}.
\code{print.rlassoEffects} prints and displays some information about fitted \code{rlassoEffect} objects.
summary.rlassoEffects summarizes information of a fitted \code{rlassoEffect} object and is described at \code{\link{summary.rlassoEffects}}.
\code{confint.rlassoEffects} extracts the confidence intervals.
\code{plot.rlassoEffects} plots the estimates with confidence intervals.
}
