% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoIV.R
\name{print.rlassoIV}
\alias{print.rlassoIV}
\alias{methods.rlassoIV}
\alias{summary.rlassoIV}
\alias{confint.rlassoIV}
\title{Methods for S3 object \code{rlassoIV}}
\usage{
\method{print}{rlassoIV}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoIV}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{confint}{rlassoIV}(object, parm, level = 0.95, ...)
}
\arguments{
\item{x}{an object of class \code{rlassoIV}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{An object of class \code{rlassoIV}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required.}
}
\description{
Objects of class \code{rlassoIV} are constructed by \code{rlassoIV}.
\code{print.rlassoIV} prints and displays some information about fitted \code{rlassoIV} objects.
\code{summary.rlassoIV} summarizes information of a fitted \code{rlassoIV} object.
\code{confint.rlassoIV} extracts the confidence intervals.
}
