% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoIVselectX.R
\name{print.rlassoIVselectX}
\alias{print.rlassoIVselectX}
\alias{methods.rlassoIVselectX}
\alias{summary.rlassoIVselectX}
\alias{confint.rlassoIVselectX}
\title{Methods for S3 object \code{rlassoIVselectX}}
\usage{
\method{print}{rlassoIVselectX}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoIVselectX}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{confint}{rlassoIVselectX}(object, parm, level = 0.95, ...)
}
\arguments{
\item{x}{an object of class \code{rlassoIVselectX}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{an object of class \code{rlassoIVselectX}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}
}
\description{
Objects of class \code{rlassoIVselectX} are constructed by \code{rlassoIVselectX}.
\code{print.rlassoIVselectX} prints and displays some information about fitted \code{rlassoIVselectX} objects.
\code{summary.rlassoIVselectX} summarizes information of a fitted \code{rlassoIVselectX} object.
\code{confint.rlassoIVselectX} extracts the confidence intervals.
}
