% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoIVselectZ.R
\name{print.rlassoIVselectZ}
\alias{print.rlassoIVselectZ}
\alias{methods.rlassoIVselectZ}
\alias{summary.rlassoIVselectZ}
\alias{confint.rlassoIVselectZ}
\title{Methods for S3 object \code{rlassoIVselectZ}}
\usage{
\method{print}{rlassoIVselectZ}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoIVselectZ}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{confint}{rlassoIVselectZ}(object, parm, level = 0.95, ...)
}
\arguments{
\item{x}{an object of class \code{rlassoIVselectZ}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{an object of class \code{rlassoIVselectZ}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required.}
}
\description{
Objects of class \code{rlassoIVselectZ} are constructed by \code{rlassoIVselectZ}. 
\code{print.rlassoIVselectZ} prints and displays some information about fitted \code{rlassoIVselectZ} objects.
\code{summary.rlassoIVselectZ} summarizes information of a fitted \code{rlassoIVselectZ} object.
\code{confint.rlassoIVselectZ} extracts the confidence intervals.
}
