% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rlassoLM.R
\name{print.rlassoLM}
\alias{confint.rlassoLM}
\alias{methods.rlassoLM}
\alias{plot.rlassoLM}
\alias{print.rlassoLM}
\alias{summary.rlassoLM}
\title{Methods for S3 object \code{rlassoLM}}
\usage{
\method{print}{rlassoLM}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoLM}(object, digits = max(3L, getOption("digits") - 3L),
  ...)

\method{confint}{rlassoLM}(object, parm, level = 0.95, joint = FALSE, ...)

\method{plot}{rlassoLM}(x, ..., var.names = T, col = "black", xlim = NULL,
  xlab = "", ylab = "", main = "", sub = "", pch = 19, cex = 1.3,
  lwd = 2, lty = 1)
}
\arguments{
\item{x}{An object of class \code{rlassoLM}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods.}

\item{object}{An object of class \code{rlassoLM}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required}

\item{joint}{logical, if joint confidence intervals should be clalculated.}
}
\description{
Objects of class \code{rlassoLM} are constructed by \code{rlassoLM.formula} or \code{rlassoLM.default}.
\code{print.rlassoLM} prints and displays some information about fitted \code{rlassoLM} objects.
\code{summary.rlassoLM} summarizes information of a fitted \code{rlassoLM} object.
\code{confint.rlassoLM} extracts the confidence intervals.
\code{plot.rlassoLM} plots the estimates.
}
\keyword{methods}
\keyword{rlassoLM}

