% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassologit.R
\name{predict.rlassologit}
\alias{predict.rlassologit}
\alias{methods.rlassologit}
\alias{print.rlassologit}
\alias{summary.rlassologit}
\alias{model.matrix.rlassologit}
\alias{model.matrix.rlassologit}
\alias{print.rlassologit}
\alias{summary.rlassologit}
\title{Methods for S3 object \code{rlassologit}}
\usage{
\method{predict}{rlassologit}(object, newdata = NULL, type = "response",
  ...)

\method{model.matrix}{rlassologit}(object, ...)

\method{print}{rlassologit}(x, all = TRUE, digits = max(3L,
  getOption("digits") - 3L), ...)

\method{summary}{rlassologit}(object, all = TRUE, digits = max(3L,
  getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{rlassologit}}

\item{newdata}{new data set for prediction}

\item{type}{type of prediction required. The default ('response) is on the scale of the response variable; the alternative 'link' is on the scale of the linear predictors.}

\item{...}{arguments passed to the print function and other methods}

\item{x}{an object of class \code{rlassologit}}

\item{all}{logical, indicates if coefficients of all variables (TRUE) should be displayed or only the non-zero ones (FALSE)}

\item{digits}{significant digits in printout}
}
\description{
Objects of class \code{rlassologit} are constructed by \code{rlassologit}.
\code{print.rlassologit} prints and displays some information about fitted \code{rlassologit} objects.
\code{summary.rlassologit} summarizes information of a fitted \code{rlassologit} object.
\code{predict.rlassologit} predicts values based on a \code{rlassologit} object.
\code{model.matrix.rlassologit} constructs the model matrix of a lasso object.
}
