% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlassologitEffect.R
\name{print.rlassologitEffect}
\alias{confint.rlassologitEffect}
\alias{methods.rlassologitEffect}
\alias{print.rlassologitEffect}
\alias{summary.rlassologitEffect}
\title{Methods for S3 object \code{rlassologitEffect}}
\usage{
\method{print}{rlassologitEffect}(x, digits = max(3L, getOption("digits") -
  3L), ...)

\method{summary}{rlassologitEffect}(object, digits = max(3L,
  getOption("digits") - 3L), ...)

\method{confint}{rlassologitEffect}(object, parm, level = 0.95,
  joint = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{rlassologitEffect}}

\item{digits}{number of significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{An object of class \code{rlassologitEffect}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{joint}{logical, if joint confidence intervals should be clalculated}
}
\description{
Objects of class \code{rlassologitEffect} are construced by \code{rlassologitEffect.default} or \code{rlassologitEffect.formula} (TBD).
\code{print.rlassologitEffect} prints and displays some information about fitted \code{rlassologitEffect} objects.
\code{summary.rlassologitEffect} summarizes information of a fitted \code{rlassologitEffect} object.
\code{confint.rlassologitEffect} extracts the confidence intervals.
}
\keyword{methods}
\keyword{rlassologitEffect}

