% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassologitEffect.R
\name{print.rlassologitEffects}
\alias{print.rlassologitEffects}
\alias{methods.rlassologitEffects}
\alias{summary.rlassologitEffects}
\alias{confint.rlassologitEffects}
\title{Methods for S3 object \code{rlassologitEffects}}
\usage{
\method{print}{rlassologitEffects}(x, digits = max(3L,
  getOption("digits") - 3L), ...)

\method{summary}{rlassologitEffects}(object, digits = max(3L,
  getOption("digits") - 3L), ...)

\method{confint}{rlassologitEffects}(object, parm, level = 0.95,
  joint = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{rlassologitEffects}}

\item{digits}{number of significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{an object of class \code{rlassologitEffects}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required.}

\item{joint}{logical, if joint confidence intervals should be clalculated}
}
\description{
Objects of class \code{rlassologitEffects} are construced by \code{rlassologitEffects} or \code{rlassologitEffect}. 
\code{print.rlassologitEffects} prints and displays some information about fitted \code{rlassologitEffect} objects.
\code{summary.rlassologitEffects} summarizes information of a fitted \code{rlassologitEffects} object.
\code{confint.rlassologitEffects} extracts the confidence intervals.
}
