% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlassoIVa.R
\name{print.rlassoManyIV}
\alias{confint.rlassoManyIV}
\alias{methods.rlassoManyIV}
\alias{print.rlassoManyIV}
\alias{summary.rlassoManyIV}
\title{Methods for S3 object \code{rlassoManyIV}}
\usage{
\method{print}{rlassoManyIV}(x, digits = max(3L, getOption("digits") - 3L),
  ...)

\method{summary}{rlassoManyIV}(object, digits = max(3L, getOption("digits") -
  3L), ...)

\method{confint}{rlassoManyIV}(object, parm, level = 0.95, ...)
}
\arguments{
\item{x}{An object of class \code{rlassoManyIV}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{An object of class \code{rlassoManyIV}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}
}
\description{
Objects of class \code{rlassoManyIV} are constructed by \code{rlassoManyIV}.
\code{print.rlassoManyIV} prints and displays some information about fitted \code{rlassoManyIV} objects.
\code{summary.rlassoManyIV} summarizes information of a fitted \code{rlassoManyIV} object.
\code{confint.rlassoManyIV} extracts the confidence intervals.
}
\keyword{methods}
\keyword{rlassoManyIV}

