% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoPDAutoDML.R
\name{print.rlassoPDAutoDML}
\alias{print.rlassoPDAutoDML}
\alias{methods.rlassoPDAutoDML}
\alias{summary.rlassoPDAutoDML}
\alias{confint.rlassoPDAutoDML}
\title{Methods for S3 object \code{rlassoPDAutoDML}}
\usage{
\method{print}{rlassoPDAutoDML}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoPDAutoDML}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{confint}{rlassoPDAutoDML}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class \code{rlassoPDAutoDML}}

\item{digits}{number of significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required.}
}
\description{
Objects of class \code{rlassoPDAutoDML} are constructed by  \code{rlassoPDAutoDML}.
\code{print.rlassoPDAutoDML} prints and displays some information about fitted \code{rlassoPDAutoDML} objects.
\code{summary.rlassoPDAutoDML} summarizes information of a fitted \code{rlassoPDAutoDML} object.
\code{confint.rlassoPDAutoDML} extracts the confidence intervals.
}
