% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassotreatment.R
\name{print.rlassoTE}
\alias{print.rlassoTE}
\alias{methods.rlassoTE}
\alias{summary.rlassoTE}
\alias{summary.rlassoTE}
\alias{confint.rlassoTE}
\title{Methods for S3 object \code{rlassoTE}}
\usage{
\method{print}{rlassoTE}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoTE}(object, digits = max(3L, getOption("digits") - 3L),
  ...)

\method{confint}{rlassoTE}(object, parm, level = 0.95, ...)
}
\arguments{
\item{x}{an object of class \code{rlassoTE}}

\item{digits}{number of significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{an object of class \code{rlassoTE}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{confidence level required.}
}
\description{
Objects of class \code{rlassoTE} are constructed by  \code{rlassoATE},  \code{rlassoATET}, \code{rlassoLATE},  \code{rlassoLATET}.
\code{print.rlassoTE} prints and displays some information about fitted \code{rlassoTE} objects.
\code{summary.rlassoTE} summarizes information of a fitted \code{rlassoTE} object.
\code{confint.rlassoTE} extracts the confidence intervals.
}
