% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rlogisticlasso.R
\name{predict.rlogisticlasso}
\alias{methods.rlogisticlasso}
\alias{model.matrix.rlogisticlasso}
\alias{predict.rlogisticlasso}
\alias{print.rlogisticlasso}
\alias{summary.rlogisticlasso}
\title{Methods for S3 object \code{rlogisticlasso}}
\usage{
\method{predict}{rlogisticlasso}(object, newdata = NULL, type = "response",
  ...)

\method{model.matrix}{rlogisticlasso}(object, ...)

\method{print}{rlogisticlasso}(x, all = TRUE, digits = max(3L,
  getOption("digits") - 3L), ...)

\method{summary}{rlogisticlasso}(object, all = TRUE, digits = max(3L,
  getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{rlogisticlasso}}

\item{newdata}{new data set for prediction}

\item{type}{the type of prediction required. The default ("response) is on the scale of the response variable; the alternative "link" is on the scale of the linear predictors.}

\item{...}{arguments passed to the print function and other methods}

\item{x}{an object of class \code{rlogisticlasso}}

\item{all}{logical, indicates if coefficients of all variables (TRUE) should be displayed or only the non-zero ones (FALSE)}

\item{digits}{significant digits in printout}
}
\description{
Objects of class \code{rlogisticlasso} are construced by \code{rlogisticlasso.formula} or \code{rlogistic.default}.
\code{print.rlogisticlasso} prints and displays some information about fitted \code{rlogisticlasso} objects.
\code{summary.rlogisticlasso} summarizes information of a fitted \code{rlasso} object.
\code{predict.rlogisticlasso} predicts values based on a lasso object.
\code{model.matrix.rlogisticlasso} constructs the model matrix of a lasso object.
}
\keyword{methods}
\keyword{rlogisticlasso}

