% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsls.R
\name{print.tsls}
\alias{print.tsls}
\alias{methods.tsls}
\alias{summary.tsls}
\title{Methods for S3 object \code{tsls}}
\usage{
\method{print}{tsls}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{tsls}(object, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{an object of class \code{tsls}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods}

\item{object}{an object of class \code{tsls}}
}
\description{
Objects of class \code{tsls} are constructed by \code{tsls}. 
\code{print.tsls} prints and displays some information about fitted \code{tsls} objects.
\code{summary.tsls} summarizes information of a fitted \code{tsls} object.
}
