% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_filter_by_mask.R
\name{occs_filter_by_mask}
\alias{occs_filter_by_mask}
\title{Function to filter ocurrence points of Species Temporal Data object (STD object)  by mask.}
\usage{
occs_filter_by_mask(this.species, mask)
}
\arguments{
\item{this.species, }{Species Temporal Data object see \code{\link[hdm]{sp_temporal_data}}.}

\item{mask}{A binary raster to filter species occurrence data.}
}
\value{
Returns un  sp.temporal.modeling object (list) con las coordenadas de los puntos de presencia, los anios las observaciones, el path de las capas temporales.
}
\description{
This function filters ocurrence points of Species Temporal Data object (STD object)  by mask.
}
