% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialROC.R
\name{PartialROC}
\alias{PartialROC}
\title{PartialROC - A modification of the function of the Barve & Barve (2016). ENMGadgets \url{https://github.com/narayanibarve}
Function PartialROC generates the area under the curve values using bootstrap method. PartialROC is a model evaluation tool, used for
continuous model outputs as compared to binary model outputs. This method is specially used for model trained using presence only data.
For more details refer DOI: 10.1016/j.ecolmodel.2007.11.008 and check ENMGadgets \url{https://github.com/narayanibarve}.}
\usage{
PartialROC(valData, PredictionFile, E = 0.05, RandomPercent, NoOfIteration)
}
\arguments{
\item{valData}{- Occurence validation data. Must have 3 columns SpName, Longitude, Latitude.}

\item{PredictionFile}{- It should be a raster class object of a continuous model output.}

\item{E}{- Amount of error admissible along the Y-axis, given the requirements and conditions of the study (by default =.05). Value should range between 0 - 1}

\item{RandomPercent}{- Occurrence points to be sampled randomly from the test data for bootstrapping.}

\item{NoOfIteration}{- Number of iteration for bootstrapping}
}
\value{
OutputFile will have 4 columns, IterationNo, AUC_at_specified_value, AUC_AT_Random, AUC_Ratio. The first row will always have 0 th interation
which is the actual Area Under the Curve without bootstrapping. And the rest of the rows contains auc ratio for all the bootstrap.
}
\description{
PartialROC - A modification of the function of the Barve & Barve (2016). ENMGadgets \url{https://github.com/narayanibarve}
Function PartialROC generates the area under the curve values using bootstrap method. PartialROC is a model evaluation tool, used for
continuous model outputs as compared to binary model outputs. This method is specially used for model trained using presence only data.
For more details refer DOI: 10.1016/j.ecolmodel.2007.11.008 and check ENMGadgets \url{https://github.com/narayanibarve}.
}
