% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML.R
\name{print.rlassoAutoDML}
\alias{print.rlassoAutoDML}
\title{prints output of rlassoAutoDML in an easy to read format}
\usage{
\method{print}{rlassoAutoDML}(obj)
}
\arguments{
\item{obj}{Object of class rlassoAutoDML}
}
\description{
prints output of rlassoAutoDML in an easy to read format
}
