% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML_helpers.R
\name{printer}
\alias{printer}
\title{prints output of rlassoAutoDML in an easy to read format}
\usage{
printer(spec1)
}
\arguments{
\item{spec1}{output of rlassoAutoDML}
}
\description{
prints output of rlassoAutoDML in an easy to read format
}
