% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{data}
\name{riboflavin}
\alias{riboflavin}
\title{Riboflavin data set}
\format{\describe{ \item{y}{Log-transformed riboflavin production rate
(original name: q_RIBFLV).} \item{x}{(Co-)variables measuring the logarithm
of the expression level of 4088 genes.} }}
\description{
Dataset of riboflavin production by Bacillus subtilis containing \eqn{n=71}
observations of \eqn{p=4088} predictors (gene expressions) and a
one-dimensional response (riboflavin production).
}
\details{
Data was made available by DSM (Switzerland). The data set was introduced in
Buehlmann et al. (2014).
}
\examples{
data(riboflavin)
}
\references{
P. Buehlmann, M. Kalisch, L- Meier (2014). High-dimensional
statistics with a view towards applications in biology. Annual Review of
Statistics and its Applications Vol. 1, 255--278.
}
\keyword{datasets}

