% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAPDAutoDML.R
\name{rlassoAPDAutoDML}
\alias{rlassoAPDAutoDML}
\alias{rlassoAPDAutoDML.default}
\title{Auto DML based on rlasso}
\usage{
rlassoAPDAutoDML(x, ...)

\method{rlassoAPDAutoDML}{default}(
  X,
  Y,
  delta = NULL,
  X.up = NULL,
  X.down = NULL,
  D_LB = 0,
  D_add = 0.2,
  debiased = TRUE,
  L = 5,
  max_iter = 10
)
}
\arguments{
\item{X}{exogenous variables}

\item{Y}{outcome variable / dependent variable}

\item{delta}{delta for partial derivative (amount of shift).}

\item{X.up}{X with upward shift for D}

\item{X.down}{X with downward shift for D}

\item{D_LB}{lower bound on D (default 0)}

\item{D_add}{value added to D (default 0.2)}

\item{debiased}{debiased (if TRUE) vs. biased (if FALSE) results.
Default is TRUE.}

\item{L}{number of folds data is split into (default 5)}

\item{max_iter}{maximum iterations of Lasso (default 10)}
}
\value{
an object of class \code{rlassoAPDAutoDML} with estimated effects,
standard errors and individual effects in the form of a \code{list}.
}
\description{
Implements AutoDML based on rigorous lasso learner (\code{rlasso}) as developed
in Chernozhukov et al. (2018b). Estimates the average partial derivative of a continuous treatment variable, \eqn{D} on an outcome variable \eqn{Y} in presence of confounding variables \eqn{X}.The algorithm is based on the double ML
cross-fitting algorithm. This method was first introduced in Chernozhukov et
al. (2018a) which required manual calculation of the Riesz representer in the
corresponding (IRM) model. In AutoDML (Chernozhukov et al., 2018b), the Riesz
representer is estimated automatically and doesn't need to be explicitly
computed.
}
\examples{
library(hdm)

}
\references{
Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E.,
Hansen, C., Newey, W. and Robins, J. (2018a), Double/debiased machine learning
for treatment and structural parameters.
The Econometrics Journal, 21: C1-C68. \doi{10.1111/ectj.12097}.

Chernozhukov, V., Newey, W.K., and Singh, R. (2018b), Automatic
Debiased Machine Learning of Causal and Structural Effects. arXiv preprint
arXiv:1809.05224 (2018).
}
\seealso{
rlasso
}
