% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML.R
\name{rlassoAutoDML}
\alias{rlassoAutoDML}
\alias{rlassoAutoDML.default}
\title{Auto DML based on rlasso}
\usage{
rlassoAutoDML(x, ...)

\method{rlassoAutoDML}{default}(
  data,
  D_LB = 0,
  D_add = 0.2,
  debiased = TRUE,
  L = 5,
  max_iter = 10,
  gamma_learner = "rlasso",
  est_type = NULL,
  prelim_est = FALSE,
  weights = NULL
)
}
\arguments{
\item{data}{Backend with preprocessed data for estimation of 
ATE (output of function DataATEAutoDML) or APD (output of DataAPDAutoDML function).}

\item{D_LB}{lower bound on D (default 0)}

\item{D_add}{value added to D (default 0.2)}

\item{debiased}{debiased (if TRUE) vs. biased (if FALSE) results.
Default is TRUE.}

\item{L}{number of folds data is split into (default 5)}

\item{max_iter}{maximum iterations of Lasso for debiasing (default 10)}

\item{est_type}{specifying the type of estimator, must either be "APD" 
for average partial derivative or "ATE" for average treatment effect.}

\item{weights}{weight vector - must be length (N-1)*T - because weight vector
for first differenced data}
}
\value{
an object of class \code{rlassoAutoDML} with estimated effects,
standard errors and individual effects in the form of a \code{list}.
}
\description{
Implements AutoDML based onlasso learners (\code{rlasso}) as developed
in Chernozhukov et al. (2022) and Klosin and Vilgalys (2022).
Estimates the average treatment effect (ATE)
of a binary treatment variable, \eqn{D} on an outcome variable \eqn{Y} in presence of confounding variables \eqn{X} or the average partial derivative (APD) of a continuous treatment variable, \eqn{D} on an outcome variable \eqn{Y}.
The algorithm is based on the double ML
cross-fitting algorithm. This method was first introduced in Chernozhukov et
al. (2018) which required manual calculation of the Riesz representer in the
corresponding (IRM) model. In AutoDML (Chernozhukov 2022), the Riesz
representer is estimated automatically and doesn't need to be explicitly
computed.
}
\examples{
library(hdm)
data <- DGP_contD(N, 20)
Xnames = colnames(data)[grep("X", colnames(data))]
backend_contD = DataAPDAutoDML(x = Xnames, d = "D", y = "y", data = data,
                              poly = 3, interactions = TRUE,
                              intercept = "intercept")
auto_apd = rlassoAutoDML(backend_contD, est_type = "APD", debiased = TRUE)
auto_apd$te

}
\references{
Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E.,
Hansen, C., Newey, W. and Robins, J. (2018a), Double/debiased machine learning
for treatment and structural parameters.
The Econometrics Journal, 21: C1-C68. \doi{10.1111/ectj.12097}.

Chernozhukov, V., Newey, W.K., and Singh, R. (2022), Automatic
Debiased Machine Learning of Causal and Structural Effects.
Econometrica, 90.3: 967-1027

Klosin, S., Vilgalys, M. (2022), Estimating Continuous
Treatment Effects in Panel Data using Machine Learning with an Agricultural
Application. arXiv preprint arXiv:2207.08789 (2022).
}
\seealso{
rlasso
}
