% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML.R
\name{rlassoAutoDML}
\alias{rlassoAutoDML}
\title{DML functional}
\usage{
rlassoAutoDML(Y, D, X, dict, D_LB = 0, D_add = 0.2, bias = FALSE, L = 5)
}
\arguments{
\item{Y}{A vector of outputs}

\item{D}{A vector of treatment values}

\item{X}{A matrix of covariates}

\item{dict}{A dictionary}
}
\value{
DML results
}
\description{
DML functional
}
\examples{
rlassoDML(Y, T, X, dict)
}
