% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlassoEffect.R
\name{rlassoEffect}
\alias{rlassoEffect}
\alias{rlassoEffect.default}
\alias{rlassoEffect.formula}
\alias{rlassoEffectone}
\title{rigorous Lasso for Linear Models: Inference}
\usage{
rlassoEffect(x, ...)

\method{rlassoEffect}{default}(x, y, index = c(1:ncol(x)), I3 = NULL, ...)

\method{rlassoEffect}{formula}(formula, data, index, I3 = NULL, ...)

rlassoEffectone(x, y, d, I3 = NULL, ...)
}
\arguments{
\item{x}{matrix of regressor variables serving as controls and potential
treatments}

\item{y}{outcome variable (vector or matrix)}

\item{index}{vector of integers, logicals or variables names indicating the position (column) of
variables (integer case), logical vector of length of the variables (TRUE or FALSE) or the variable names of \code{x} which should be used for inference / as treatment variables.}

\item{I3}{logical vector with the same length as the number of variables in \code{x};
indicates if variables (TRUE) should be included in any case to the model and they are exempt from selection. These variables should not be included in the \code{index}; hence the intersection with \code{index} must be the empty set.}

\item{formula}{an element of class \code{formula}}

\item{data}{data.frame in connection with \code{formula}}

\item{d}{variable for which inference is conducted (treatment variable)}

\item{\dots}{parameters passed to the \code{\link{rlasso}} function.}
}
\value{
The function returns an object of class \code{rlassoEffect} with the following entries: \item{coefficients}{vector with estimated
values of the coefficients for each selected variable} \item{se}{standard error (vector)}
\item{t}{t-statistic} \item{pval}{p-value} \item{samplesize}{sample size of the data set} \item{index}{index of the variables for which inference is performed}
}
\description{
Estimation and inference of selected (low-dimensional) coefficients in a high-dimensional linear model.
}
\details{
The functions estimates selected (low-dimensional) coefficients in a high-dimensional linear model.
An application is e.g. estimation of a treatment effect \eqn{\alpha_0} in a
setting of high-dimensional controls. The so-called post-double-selection
estimator is implemented. The idea is to select variables by regression of
the outcome variable on the control variables and the treatment variable on
the control variables. The final estimation is done by a regression of the
outcome on the treatment effect and the union of the selected variables in
the first two steps. The resulting estimator for \eqn{\alpha_0} is normal
distributed which allows inference on the treatment effect. It presents a wrap function for \code{rlassoEffectone}
which does inference for a single variable.
}
\examples{
library(hdm)
## DGP
n <- 250
p <- 100
px <- 10
X <- matrix(rnorm(n*p), ncol=p)
beta <- c(rep(2,px), rep(0,p-px))
intercept <- 1
y <- intercept + X \%*\% beta + rnorm(n)
## fit rlassoEffect object with inference on three variables
rlassoEffect.reg <- rlassoEffect(x=X, y=y, index=c(1,7,20))
## methods
summary(rlassoEffect.reg)
confint(rlassoEffect.reg, level=0.9)
}
\references{
A. Belloni, V. Chernozhukov, C. Hansen (2014). Inference on
treatment effects after selection among high-dimensional controls. The
Review of Economic Studies 81(2), 608-650.
}
\keyword{Estimation}
\keyword{High-dimensional}
\keyword{Inference}
\keyword{Treatment}
\keyword{controls}
\keyword{effect}

