% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rlassoLM.R
\name{rlassoLM}
\alias{rlassoLM}
\alias{rlassoLM.default}
\alias{rlassoLM.formula}
\alias{rlassoLMone}
\title{rigorous Lasso for Linear Models: Inference}
\usage{
rlassoLM(x, ...)

\method{rlassoLM}{default}(x, y, index = c(1:ncol(x)), I3 = NULL, ...)

\method{rlassoLM}{formula}(formula, data, index, I3 = NULL, ...)

rlassoLMone(x, y, d, I3 = NULL, ...)
}
\arguments{
\item{x}{matrix of regressor variables serving as controls and potential
treatments}

\item{y}{outcome variable (vector or matrix)}

\item{index}{vector of integers, logicals or variables names indicating the position (column) of
variables (integer case), logical yes or no (TRUE or FALSE) or the variable names of x which should be used for inference / as treatment variables.}

\item{I3}{logical vector with same length as the number of controls;
indicates if variables (TRUE) should be included in any case.}

\item{formula}{an element of class \code{formula}}

\item{data}{data.frame in connection with \code{formula}}

\item{d}{variable for which inference is conducted (treatment variable)}

\item{\dots}{parameters passed to the \code{\link{rlasso}} function.}
}
\value{
The function returns an object of class \code{rlassoLM} with the following entries: \item{coefficients}{vector with estimated
values of the coefficients for each selected variable} \item{se}{standard error (vector)}
\item{t}{t-statistic} \item{pval}{p-value} \item{samplesize}{sample size of the data set} \item{I}{index of variables of the union of the lasso regressions}
}
\description{
Estimation and inference of selected (low-dimensional) coefficients in a high-dimensional linear model.
}
\details{
The functions estimates selected (low-dimensional) coefficients in a high-dimensional linear model.
An application is e.g. estimation of a treatment effect \eqn{\alpha_0} in a
setting of high-dimensional controls. The so-called post-double-selection
estimator is implemented. The idea is to select variable by regression of
the outcome variable on the control variables and the treatment variable on
the control variables. The final estimation is done by a regression of the
outcome on the treatment effect and the union of the selected variables in
the first two steps. The resulting estimator for \eqn{\alpha_0} is normal
distributed which allows inference on the treatment effect.
}
\references{
A. Belloni, V. Chernozhukov, C. Hansen (2014). Inference on
treatment effects after selection among high-dimensional controls. The
Review of Economic Studies 81(2), 608--650.
}
\keyword{Estimation}
\keyword{High-dimensional}
\keyword{Inference}
\keyword{Treatment}
\keyword{controls}
\keyword{effect}

