% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rlassologitEffect.R
\name{rlassologitEffect}
\alias{rlassologitEffect}
\alias{rlassologitEffect.default}
\alias{rlassologitEffectone}
\title{rigorous Lasso for Logistic Models: Inference}
\usage{
rlassologitEffect(x, ...)

\method{rlassologitEffect}{default}(x, y, index = c(1:ncol(x)), I3 = NULL,
  ...)

rlassologitEffectone(x, y, d, I3 = NULL)
}
\arguments{
\item{x}{matrix of regressor variables serving as controls and potential
treatments}

\item{y}{outcome variable}

\item{index}{vector of integers indicating the position (column) of
variables of x which should be used as treatment variables.}

\item{I3}{logical vector with same length as the number of controls;
indicates if variables (TRUE) should be included in any case.}

\item{d}{variable for which inference is conducted (treatment variable)}

\item{\dots}{parameters passed to the lasso functions used for the double selection procedure.}

\item{\dots}{additional paramters}
}
\value{
The function returns an object of class \code{rlassologitEffect} with the following entries: \item{coefficients}{estimated
value of the coefficients} \item{se}{standard error}
\item{t}{t-statistic} \item{pval}{p-value} \item{samplesize}{sample size of the data set} \item{I}{index of variables of the union of the lasso regressions}
}
\description{
The functions estimates selected (low-dimensional) coefficients in a high-dimensional logistic model.
}
\details{
The functions estimates selected (low-dimensional) coefficients in a high-dimensional logistic model.
An application is e.g. estimation of a treatment effect \eqn{\alpha_0} in a
setting of high-dimensional controls. The function is a wrap function for \code{rlassologitEffectone} which does inference for only one variable (d).
}
\references{
A. Belloni, V. Chernozhukov, Y. Wei (2013). Honest confidence regions for a regression parameter in logistic regression with a loarge number of controls.
cemmap working paper CWP67/13.
}
\keyword{Estimation}
\keyword{Inference}
\keyword{Lasso}
\keyword{Logistic}

