% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlassoIVa.R
\name{rlassoManyIV}
\alias{rlassoManyIV}
\alias{rlassoManyIVmult}
\alias{rlassolassoManyIVmult}
\title{Post-Selection and Post-Regularization Inference in Linear Models with Many
Controls and Instruments}
\usage{
rlassoManyIV(x, z, y, d, ...)

rlassoManyIVmult(x, z, y, d, ...)
}
\arguments{
\item{x}{matrix of exogenous variables}

\item{z}{matrix of instrumental variables}

\item{y}{outcome / dependent variable (vector or matrix)}

\item{d}{endogenous variable}

\item{\dots}{arguments passed to the function \code{rlasso}}
}
\value{
An object of class \code{rlassoManyIV} containing at least the following
components: \item{coefficients}{estimated parameter value}
\item{se}{variance-covariance matrix}
}
\description{
The function estimates a treatment effect in a setting with very many
controls and very many instruments (even larger than the sample size).
}
\details{
The implementation follows the procedure described in Chernozhukov et al.
(2015) and is built on "triple selection" to achieve an orthogonal moment
function. The function returns an object of S3 class \code{rlassoManyIV}.
Moreover, it is wrap function for the case that selection should be done only with the instruments Z (\code{rlassoIVselectZ}) or with
the control variables X (\code{rlassoIVselectX}).
}
\references{
V. Chernozhukov, C. Hansen, M. Spindler (2015). Post-selection
and post-regularization inference in linear models with many controls and
instruments. American Economic Review: Paper & Proceedings 105(5), 1--7.
}
\keyword{Instrumental}
\keyword{Lasso}
\keyword{Many}
\keyword{Variable}
\keyword{and}
\keyword{controls}
\keyword{instruments}

