% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rlogisticlasso.R
\name{rlogisticlasso}
\alias{rlogisticlasso}
\alias{rlogisticlasso.default}
\alias{rlogisticlasso.formula}
\title{rlogisticlasso: Function for logistic Lasso estimation}
\usage{
rlogisticlasso(x, ...)

\method{rlogisticlasso}{default}(x, y, post = TRUE, intercept = TRUE,
  normalize = TRUE, penalty = list(lambda = NULL, c = 1.1, gamma =
  0.1/log(n)), control = list(threshold = NULL), ...)

\method{rlogisticlasso}{formula}(formula, data, post = TRUE,
  intercept = TRUE, normalize = TRUE, penalty = list(c = 1.1, gamma =
  0.1/log(n)), control = list(threshold = NULL), ...)
}
\arguments{
\item{x}{regressors (matrix)}

\item{...}{further parameters passed to glmnet}

\item{y}{dependent variable (vector or matrix)}

\item{post}{logical. If \code{TRUE}, post-lasso estimation is conducted.}

\item{intercept}{logical. If \code{TRUE}, intercept is included which is not
penalized.}

\item{normalize}{logical. If \code{TRUE}, design matrix \code{x} is scaled.}

\item{penalty}{list with options for the calculation of the penalty.  \code{c} and \code{gamma} constants for the penalty.}

\item{control}{list with control values.
\code{threshold} is applied to the final estimated lasso
coefficients. Absolute values below the threshold are set to zero.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted in the form
\code{y~x}}

\item{data}{an optional data frame, list or environment (or object coercible}
}
\value{
\code{rlogisticlasso} returns an object of class
\code{rlogisticlasso} An object of class \code{rlogisticlasso} is a list
containing at least the following components: \item{coefficients}{parameter
estimates (without intercept)} \item{a0}{value of intercept} \item{index}{index of selected variables (logicals)}
\item{lambda0}{penalty term}
\item{residuals}{residuals}
\item{sigma}{root of the variance of the residuals}
\item{call}{function call}
\item{options}{options}
}
\description{
The function estimates the coefficients of a logistic Lasso regression with
data-driven penalty. The method of the data-driven penalty can be chosen.
The object which is returned is of the S3 class \code{rlogisticlasso}
}
\details{
The function estimates the coefficients of a Logistic Lasso regression with
data-driven penalty. The method of the data-driven penalty can be chosen
("X-dependent", "X-independent", "standard", "none"). The description of the
methods is given in \code{\link{rlasso}}, with the minor difference that here
also for the "X-dependent" and "X-independent" method data-driven loadings
are employed (instead of an iteratively estimated variance factor). The
option \code{post=TRUE} conducts post-lasso estimation, i.e. a refit of the
model with the selected variables.
}
\references{
TBD
}
\keyword{lasso}
\keyword{logistic}
\keyword{regression}

