% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML_helpers.R
\name{RMD_stable}
\alias{RMD_stable}
\title{Estimation of Riesz Representer}
\usage{
RMD_stable(
  Y,
  X,
  M,
  p,
  D_LB = 0,
  D_add = 0.2,
  max_iter = 10,
  c = 0.5,
  gamma = 0.1,
  tol = 1e-06,
  prelim_quantities = NULL
)
}
\arguments{
\item{D_LB}{lower bound on D (default 0)}

\item{D_add}{value added to D (default 0.2)}

\item{max_iter}{maximum iterations of Lasso for debiasing (default 10)}

\item{c}{parameter to tune lambda (default 0.5)}

\item{gamma}{parameter to tune lambda (default 0.1)}

\item{tol}{minimum improvement to continue looping (default 1e-6)}
}
\description{
Implements estimation of the Riesz representer \eqn{\alpha}.
}
\details{
TODO: Finalize documentation
}
