% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoEffects.R
\name{summary.rlassoEffects}
\alias{summary.rlassoEffects}
\alias{print.summary.rlassoEffects}
\title{Summarizing rlassoEffects fits}
\usage{
\method{summary}{rlassoEffects}(object, ...)

\method{print}{summary.rlassoEffects}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{rlassoEffects}, usually a result of a call to \code{rlassoEffects}}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{summary.rlassoEffects}, usually a result of a call or \code{summary.rlassoEffects}}

\item{digits}{the number of significant digits to use when printing.}
}
\description{
Summary method for class \code{rlassoEffects}
}
\details{
Summary of objects of class \code{rlassoEffects}
}
