% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_projection.R
\name{temporal_projection}
\alias{temporal_projection}
\title{Temporal projection of the best model.
The function projects the niche model for the d}
\usage{
temporal_projection(this.species, save_dir, sp_mask, sp_name = "",
  crs_model = NULL)
}
\arguments{
\item{this.species}{a "sp.temp.best.model" object see \code{\link[hdm]{find_best_model}}.}

\item{save_dir}{Folder where the projected model results will be saved.}

\item{sp_mask}{A mask to crop the projected area of distribution.}

\item{sp_name}{The name of the modeled species.}

\item{crs_model}{The crs of the projection layers.}
}
\value{
The output is contained in main directory called "temporal_modeling", it is arranged in subdirectories corresponding to each of the projected years. Each subdirectory has:
- Ellipsoid model plot. Only If the number of dimensions to model the niche where 2 or 3.
- Binary model of distribution.
- A continuos suitability raster.
Finally a directory called niche_comparations_results, which has barplots of the percent of area/suitabilty gained (or lost) in each year compared to \code{t_0}.
}
\description{
Temporal projection of the best model.
The function projects the niche model for the d
}
