% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoDebiased_backends.R
\name{DataATEAutoDML}
\alias{DataATEAutoDML}
\title{Backend for rlassoATEAutoDML}
\usage{
DataATEAutoDML(x, d, y, data, dict = NULL)
}
\arguments{
\item{x}{character specifying exogenous variables}

\item{d}{character specifying treatment variable (binary)}

\item{y}{character specifying outcome variable / dependent variable}

\item{data}{data frame or data.table}

\item{dict}{a dictionary (function)
function of (D,X) with D being a numeric and X a matrix, which returns a numeric vector. Default is \code{dict <- function(d, x) return(c(1, d, x))}. The dictionary will be applied on an individual level, i.e., being executed for each i in 1, ..., n.}
}
\value{
an object of class \code{DataAutoDebiasedATE} with the following entries
\itemize{
  \item{\code{Y} - }{Y as passed as input}
  \item{X - }{Processed X (dictionary applied to input data)}
  \item{dict - }{dictionary}
  
  \item{"D_is_binary" - }{logical indicating whether the treatment variable is binary}
  \item{"M" - }{Matrix M as of ...}
  \item{"G" - }{Matrix G as of ...}
}
}
\description{
Data backend for auto debiased ATE. Implements the preprocessing steps required for estimation of the ATE using auto debiased ATE.
}
