% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{data}
\name{EminentDomain}
\alias{EminentDomain}
\title{Eminent Domain data set}
\format{\describe{
\item{y}{economic outcome variable, here the CS Home Price Index}
\item{x}{set of exogenous variables}
\item{d}{eminent domain decisions}
\item{z}{set of potential instruments}
}}
\description{
Dataset on judicial eminent domain decisions.
}
\details{
Data set was analyzed in Belloni et al. (2012).  They estimate the effect of judicial eminent domain decisions on economic outcomes with instrumental variable (IV) in a setting high a large set of potential IVs.
A detailed decription of the data can be found at
\url{https://www.econometricsociety.org/publications/econometrica/2012/11/01/sparse-models-and-methods-optimal-instruments-application}
}
\examples{
data(EminentDomain)
}
\references{
D. Belloni, D. Chen, V. Chernozhukov and C. Hansen (2012).
Sparse models and methods for optimal instruments with an application to
eminent domain. \emph{Econometrica} 80 (6), 2369--2429.
}
\keyword{datasets}

