\name{hdm-package}
\alias{hdm-package}
\alias{hdm}
\docType{package}
\title{hdm: High-Dimensional Metrics
}
\description{
This packages implements high-dimensional methods with a focus on Econometrics.
}
\details{
\tabular{ll}{
Package: \tab hdm\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-05-25\cr
License: \tab GPL-3\cr
}
The package includes functions for fitting a heteroskedastic robust Lasso regression and for instrumental variable (IV) and treatment effect estimation in a high-dimensional setting. Moreover, the methods enable valid post-selection inference.
}
\author{
Victor Chernozhukov, Christian Hansen, Martin Spindler\cr

Maintainer: Martin Spindler <spindler@mea.mpisoc.mpg.de>

}
\references{TBD
}
\keyword{package}
\keyword{Lasso}
\keyword{Instrumental Variables}
\keyword{Endogeneity}
\keyword{Microeconometrics}
\keyword{Program Evaluation}
\keyword{treatment effects}
