% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlasso.R
\name{lambdaCalculation}
\alias{lambdaCalculation}
\title{Function for Calculation of the penalty parameter}
\usage{
lambdaCalculation(penalty = list(method = "standard", lambda.start = NULL, c =
  1.1, gamma = 0.1), y = NULL, x = NULL)
}
\arguments{
\item{penalty}{list with options for the calculation of the penalty.  \code{c} and \code{gamma} constants for the penalty (all methods except "CV"), \code{method} method for penalty choice ("X-dependent",
"X-independent", "standard", "none", "CV"), \code{numfolds} number of folds
for the "cross validation" method, \code{numSim} number of simulations for
the "X-dependent" method, \code{lambda.start} initial penalization value, compulsory for method "none"}

\item{y}{residual which is used for calculation of the variance or the data-dependent loadings}

\item{x}{matrix of regressor variables}
}
\value{
The functions returns a list with the penalty \code{lambda} which is the product of \code{lambda0} and \code{Ups0}. \code{Ups0}
denotes either the variance or the data-dependent loadings for the regressors. \code{method} gives the selected method for the calculation.
}
\description{
This function implements different methods for calculation of the parameter lambda. Further details can be found under \link{rlasso}.
}

