% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlassoEffect.R
\name{print.rlassoEffects}
\alias{confint.rlassoEffects}
\alias{methods.rlassoEffects}
\alias{plot.rlassoEffects}
\alias{print.rlassoEffects}
\alias{summary.rlassoEffects}
\title{Methods for S3 object \code{rlassoEffects}}
\usage{
\method{print}{rlassoEffects}(x, digits = max(3L, getOption("digits") - 3L),
  ...)

\method{summary}{rlassoEffects}(object, digits = max(3L, getOption("digits") -
  3L), ...)

\method{confint}{rlassoEffects}(object, parm, level = 0.95, joint = FALSE,
  ...)

\method{plot}{rlassoEffects}(x, main = "", xlab = "coef", ylab = "",
  xlim = NULL, col = "black", ...)
}
\arguments{
\item{x}{An object of class \code{rlassoEffects}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods.}

\item{object}{An object of class \code{rlassoEffects}}

\item{parm}{a specification of which parameters are to be given confidence intervals among the variables for which inference was done, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required}

\item{joint}{logical, if \code{TRUE} joint confidence intervals are calculated.}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{xlim}{vector of length two giving lower and upper bound of x axis}

\item{col}{color of lines of the graph}
}
\description{
Objects of class \code{rlassoEffects} are constructed by \code{rlassoEffects.formula} or \code{rlassoEffects.default}.
\code{print.rlassoEffects} prints and displays some information about fitted \code{rlassoEffect} objects.
\code{summary.rlassoEffects} summarizes information of a fitted \code{rlassoEffect} object.
\code{confint.rlassoEffects} extracts the confidence intervals.
\code{plot.rlassoEffects} plots the estimates with confidence intervals.
}
\keyword{methods}
\keyword{rlassoEffects}

