% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlassoLM.R
\name{print.rlassoLM}
\alias{confint.rlassoLM}
\alias{methods.rlassoLM}
\alias{plot.rlassoLM}
\alias{print.rlassoLM}
\alias{summary.rlassoLM}
\title{Methods for S3 object \code{rlassoLM}}
\usage{
\method{print}{rlassoLM}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{summary}{rlassoLM}(object, digits = max(3L, getOption("digits") - 3L),
  ...)

\method{confint}{rlassoLM}(object, parm, level = 0.95, joint = FALSE, ...)

\method{plot}{rlassoLM}(x, main = "", xlab = "coef", ylab = "",
  xlim = NULL, col = "black", ...)
}
\arguments{
\item{x}{An object of class \code{rlassoLM}}

\item{digits}{significant digits in printout}

\item{...}{arguments passed to the print function and other methods.}

\item{object}{An object of class \code{rlassoLM}}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required}

\item{joint}{logical, if \code{TRUE} joint confidence intervals are clalculated.}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{xlim}{vector of length two giving lower and upper bound of x axis}

\item{col}{color of lines of the graph}
}
\description{
Objects of class \code{rlassoLM} are constructed by \code{rlassoLM.formula} or \code{rlassoLM.default}.
\code{print.rlassoLM} prints and displays some information about fitted \code{rlassoLM} objects.
\code{summary.rlassoLM} summarizes information of a fitted \code{rlassoLM} object.
\code{confint.rlassoLM} extracts the confidence intervals.
\code{plot.rlassoLM} plots the estimates with confidence intervals.
}
\keyword{methods}
\keyword{rlassoLM}

