% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlasso.R
\name{rlasso}
\alias{rlasso}
\alias{rlasso.default}
\alias{rlasso.formula}
\title{rlasso: Function for Lasso estimation under homoskedastic and heteroskeadstic non-Gaussian
disturbances}
\usage{
rlasso(x, ...)

\method{rlasso}{default}(x, y, post = TRUE, intercept = TRUE,
  normalize = TRUE, penalty = list(homoscedastic = FALSE, X.design =
  "independent", lambda.start = NULL, c = 1.1, gamma = 0.1),
  control = list(numIter = 15, tol = 10^-5, threshold = NULL), ...)

\method{rlasso}{formula}(formula, data, post = TRUE, intercept = TRUE,
  normalize = TRUE, penalty = list(homoscedastic = FALSE, X.design =
  "independent", lambda.start = NULL, c = 1.1, gamma = 0.1),
  control = list(numIter = 15, tol = 10^-5, threshold = NULL), ...)
}
\arguments{
\item{x}{regressors (matrix or data.frame)}

\item{...}{further arguments (only for consistent defintion of methods)}

\item{y}{dependent variable (vector or matrix)}

\item{post}{logical. If \code{TRUE}, post-lasso estimation is conducted.}

\item{intercept}{logical. If \code{TRUE}, intercept is included which is not
penalized.}

\item{normalize}{logical. If \code{TRUE}, design matrix \code{x} is scaled.}

\item{penalty}{list with options for the calculation of the penalty.
\itemize{
\item{\code{c} and \code{gama}}{constants for the penalty with default \code{c=1.1} and \code{gamma=0.1}}
\item{\code{homoscedastic}}{logical, if homoscedastic errors are considered (default \code{FALSE}). Option \code{none} is described below.}
\item{\code{X.design}}{if \code{independent} or \code{dependent} design matrix \code{X}}
\item{\code{numSim}} number of simulations for the dependent methods
\item{\code{lambda.start}}{initial penalization value, compulsory for method "none"}
}}

\item{control}{list with control values.
\code{numIter} number of iterations for the algorithm for
the estimation of the variance and data-driven penalty, ie. loadings,
\code{tol} tolerance for improvement of the estimated variances.
\code{threshold} is applied to the final estimated lasso
coefficients. Absolute values below the threshold are set to zero.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted in the form
\code{y~x}}

\item{data}{an optional data frame, list or environment (or object coercible
by as.data.frame to a data frame) containing the variables in the model. If
not found in data, the variables are taken from environment(formula),
typically the environment from which \code{rlasso} is called.}
}
\value{
\code{rlasso} returns an object of class \code{rlasso} An object of
class "rlasso" is a list containing at least the following components:
\item{coefficients}{parameter estimates (named vector of coefficients without intercept)}
\item{intercept.value}{value of the intercept}
\item{index}{index of selected
variables (logical vector)} \item{lambda}{data-driven penalty term for each
variable, product of lambda0 (the penalization parameter) and the loadings}
\item{lambda0}{penalty term} \item{loadings}{loading for each regressor}
\item{residuals}{residuals, response minus fitted values} \item{sigma}{root of the variance of
the residuals} \item{iter}{number of iterations} \item{call}{function call}
\item{options}{options}
}
\description{
The function estimates the coefficients of a Lasso regression with
data-driven penalty under homoskedasticity and heteroskedasticity with non-Gaussian noise. The
method of the data-driven penalty can be chosen. The object which is
returned is of the S3 class \code{rlasso}.
}
\details{
The function estimates the coefficients of a Lasso regression with
data-driven penalty under homoskedasticity / heteroskedasticity and non-Gaussian noise. The options \code{homoscedatic} is a logical with \code{FALSE} by default.
Moreover, for the calculation of the penalty parameter it can be chosen, if the design matrix (\code{X.design}) is \code{independent} (default option) or \code{dependent}.
A \emph{special} option is to set \code{homoscedastic} to \code{none} and to supply a values \code{lambda.start}. Then this values is used as penalty parameter with independet design to weight the regressors.
For details of the
implementation of the Algorithm for estimation of the data-driven penalty,
in particular the regressor-independent loadings, we refer to Appendix A in
Belloni et al.~(2012). When the option "none" is chosen (together with
\code{lambda.start}), lambda is set to \code{lambda.start} and the
regressor-independent loadings are used. The options "X-dependent" and
"X-independent" under homoscedasticity are described in Belloni et al.~(2013).
\code{lambda.start} can be component-specific. When used with one of the
other option, the values are used as starting values.

The option \code{post=TRUE} conducts post-lasso estimation, i.e. a refit of
the model with the selected variables.
}
\references{
A. Belloni, D. Chen, V. Chernozhukov and C. Hansen (2012).
Sparse models and methods for optimal instruments with an application to
eminent domain. \emph{Econometrica} 80 (6), 2369-2429.

A. Belloni, V. Chernozhukov and C. Hansen (2013). Inference for
high-dimensional sparse econometric models. In Advances in Economics and
Econometrics: 10th World Congress, Vol. 3: Econometrics, Cambirdge
University Press: Cambridge, 245-295.
}
\keyword{Lasso}
\keyword{data-driven}
\keyword{heteroscedasticity}
\keyword{non-Gaussian}
\keyword{penalty}

