% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlasso.R
\name{rlasso}
\alias{rlasso}
\alias{rlasso.default}
\alias{rlasso.formula}
\title{rlasso: Function for Lasso estimation under homoskedastic and heteroskeadstic non-Gaussian
disturbances}
\usage{
rlasso(x, ...)

\method{rlasso}{default}(x, y, post = TRUE, intercept = TRUE,
  normalize = TRUE, penalty = list(method = "standard", lambda.start = NULL,
  c = 1.1, gamma = 0.1), control = list(numIter = 15, tol = 10^-5, threshold =
  NULL), ...)

\method{rlasso}{formula}(formula, data, post = TRUE, intercept = TRUE,
  normalize = TRUE, penalty = list(method = "standard", lambda.start = NULL,
  c = 1.1, gamma = 0.1), control = list(numIter = 15, tol = 10^-5, threshold =
  NULL), ...)
}
\arguments{
\item{x}{regressors (matrix or data.frame)}

\item{...}{further arguments (only for consistent defintion of methods)}

\item{y}{dependent variable (vector or matrix)}

\item{post}{logical. If \code{TRUE}, post-lasso estimation is conducted.}

\item{intercept}{logical. If \code{TRUE}, intercept is included which is not
penalized.}

\item{normalize}{logical. If \code{TRUE}, design matrix \code{x} is scaled.}

\item{penalty}{list with options for the calculation of the penalty.  \code{c} and \code{gamma} constants for the penalty (for all methods except "CV") with default \code{c=1} and \code{gamma=0.1}, \code{method} method for penalty choice ("X-dependent",
"X-independent", "standard", "none", "CV"), \code{numfolds} number of folds
for the "cross validation" method, \code{numSim} number of simulations for
the "X-dependent" method, \code{lambda.start} initial penalization value, compulsory for method "none"}

\item{control}{list with control values.
\code{numIter} number of iterations for the algorithm for
the estimation of the variance and data-driven penalty, ie. loadings,
\code{tol} tolerance for improvement of the estimated variances.
\code{threshold} is applied to the final estimated lasso
coefficients. Absolute values below the threshold are set to zero.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted in the form
\code{y~x}}

\item{data}{an optional data frame, list or environment (or object coercible
by as.data.frame to a data frame) containing the variables in the model. If
not found in data, the variables are taken from environment(formula),
typically the environment from which \code{rlasso} is called.}
}
\value{
\code{rlasso} returns an object of class \code{rlasso} An object of
class "rlasso" is a list containing at least the following components:
\item{coefficients}{parameter estimates (named vector of coefficients without intercept)}
\item{intercept.value}{value of the intercept}
\item{index}{index of selected
variables (logical vector)} \item{lambda}{data-driven penalty term for each
variable, product of lambda0 (the penalization parameter) and the loadings}
\item{lambda0}{penalty term} \item{loadings}{loading for each regressor}
\item{residuals}{residuals, response minus fitted values} \item{sigma}{root of the variance of
the residuals} \item{iter}{number of iterations} \item{call}{function call}
\item{options}{options}
}
\description{
The function estimates the coefficients of a Lasso regression with
data-driven penalty under homoskedasticity and heteroskedasticity with non-Gaussian noise. The
method of the data-driven penalty can be chosen. The object which is
returned is of the S3 class \code{rlasso}.
}
\details{
The function estimates the coefficients of a Lasso regression with
data-driven penalty under heteroskedasticity and non-Gaussian noise. The
method of the data-driven penalty can be chosen ("X-dependent",
"X-independent", "standard", "none", "cv"). For details of the
implementation of the Algorithm for estimation of the data-driven penalty,
in particular the regressor-dependent loadings, we refer to Appendix A in
Belloni et al.~(2012). When the option "none" is chosen (together with
\code{lambda.start}), lambda is set to \code{lambda.start} and the
regressor-dependent loadings are used. The options "X-dependent" and
"X-independent" are described in Belloni et al.~(2013). The method "cv" does
cross validation to determine the "optimal" value for lambda (with
regressor-dependent loadings). The option \code{nfolds} gives the number of
folds used for cross validation. The method "standard", which is recommended
and the default option, employs data-dependent loadings and sets
\eqn{\lambda = 2*c*\sqrt{n}*\sqrt{2*\log(2*p*\log(n)/\gamma)}}.
\code{lambda.start} can be component-specific. When used with one of the
other option, the values are used as starting values.

The option \code{post=TRUE} conducts post-lasso estimation, i.e. a refit of
the model with the selected variables.
}
\examples{
## DGP
library(hdm)
library(MASS)
n <- 250
p <- 100
px <- 10
X <- matrix(rnorm(n*p), ncol=p)
beta <- c(rep(2,px), rep(0,p-px))
y <- 1 + X\%*\%beta + rnorm(n)
## Lasso estimation
lasso.reg <- rlasso(x=X, y=y, post=TRUE, intercept=TRUE)
# Methods for Lasso
print(lasso.reg, all=FALSE)
summary(lasso.reg, all=FALSE)
yhat <- predict(lasso.reg)
Xnew <- matrix(rnorm(n*p), ncol=p)
yhat.new <- predict(lasso.reg, newdata=Xnew)
}
\references{
A. Belloni, D. Chen, V. Chernozhukov and C. Hansen (2012).
Sparse models and methods for optimal instruments with an application to
eminent domain. \emph{Econometrica} 80 (6), 2369-2429.

A. Belloni, V. Chernozhukov and C. Hansen (2013). Inference for
high-dimensional sparse econometric models. In Advances in Economics and
Econometrics: 10th World Congress, Vol. 3: Econometrics, Cambirdge
University Press: Cambridge, 245-295.
}
\keyword{Lasso}
\keyword{data-driven}
\keyword{heteroscedasticity}
\keyword{non-Gaussian}
\keyword{penalty}

