% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML.R
\name{rlassoAutoDML}
\alias{rlassoAutoDML}
\alias{rlassoAutoDML.default}
\title{Auto DML based on rlasso}
\usage{
rlassoAutoDML(x, ...)

\method{rlassoAutoDML}{default}(
  X,
  D,
  Y,
  dict = NULL,
  D_LB = 0,
  D_add = 0.2,
  debiased = TRUE,
  L = 5,
  max_iter = 10,
  p0 = NULL
)
}
\arguments{
\item{X}{exogenous variables}

\item{D}{treatment variable (binary)}

\item{Y}{outcome variable / dependent variable}

\item{dict}{a dictionary
function of (d,z) that maps to a vector. Default is (1,d,z)}

\item{D_LB}{lower bound on D (default 0)}

\item{D_add}{value added to D (default 0.2)}

\item{debiased}{debiased (if TRUE) vs. biased (if FALSE) results.
Default is TRUE.}

\item{L}{number of folds data is split into (default 5)}

\item{max_iter}{maximum iterations of Lasso (default 10)}

\item{p0}{initial dimension used in preliminary estimation step. By default, a rule of thumb is applied:
If \eqn{p \le 60}, p0 = p/4, else p0 = p/40.}
}
\value{
an object of class \code{rlassoAutoDML} with estimated effects,
standard errors and individual effects in the form of a \code{list}.
}
\description{
Implements AutoDML based on rigorous lasso learner (\code{rlasso}) as developed
in Chernozhukov et al. (2018b). The algorithm is based on the double ML
cross-fitting algorithm. This method was first introduced in Chernozhukov et
al. (2018a) which required manual calculation of the Riesz representer in the
corresponding (IRM) model. In AutoDML (Chernozhukov et al., 2018b), the Riesz
representer is estimated automatically and doesn't need to be explicitly
computed.
}
\examples{
library(hdm)
library(DoubleML)
data <- make_irm_data(theta = 2, return_type = "matrix")
Y <- data$y
D <- data$d
X <- data$X

# Define a dictionary
b2 <- function(d,z){
  return(c(1,d,z,d*z))
}
lasso.auto <- rlassoAutoDML(X, D, Y, dict = b2)
summary(lasso.auto)
print(lasso.auto)
confint(lasso.auto)

}
\references{
Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E.,
Hansen, C., Newey, W. and Robins, J. (2018a), Double/debiased machine learning
for treatment and structural parameters.
The Econometrics Journal, 21: C1-C68. \doi{10.1111/ectj.12097}.

Chernozhukov, V., Newey, W.K., and Singh, R. (2018b), Automatic
Debiased Machine Learning of Causal and Structural Effects. arXiv preprint
arXiv:1809.05224 (2018).
}
\seealso{
rlasso
}
