% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML.R
\name{rlassoAutoDML}
\alias{rlassoAutoDML}
\title{Auto DML based on rlasso}
\usage{
rlassoAutoDML(
  Y,
  D,
  X,
  dict = NULL,
  D_LB = 0,
  D_add = 0.2,
  bias = FALSE,
  L = 5,
  max_iter = 10
)
}
\arguments{
\item{Y}{A vector of outputs}

\item{D}{A vector of treatment values}

\item{X}{A matrix of covariates}

\item{dict}{A dictionary
function of (d,z) that maps to a vector
default is (1,d,z)}

\item{D_LB}{Lower bound on D (default 0)}

\item{D_add}{value added to D (default 0.2)}

\item{bias}{debiased vs. biased results}

\item{L}{number of folds data is split into (default 5)}

\item{max_iter}{maximum iterations of Lasso (default 10)}
}
\value{
list with average treatment effect and standard error
}
\description{
Implements the Double ML algorithm introduced in for estimating causal effects
This method was first introdued in https://arxiv.org/abs/1608.00060 which required
manual calculation of the Riesz representer. In Auto DML (from https://arxiv.org/abs/1809.05224)
the Riesz representer is estimated automatically and doesn't need to be explicitly 
computed.
}
\details{
This implementation
}
\examples{
# data = simulate_data(500)
#
# Y = data[[1]]
# D = data[[2]]
# X = data[[3]]

# rlassoAutoDML(Y,D,X,dict = b2)
# rlassoAutoDML(Y, T, X)
}
