% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML_helpers.R
\name{RMD_stable}
\alias{RMD_stable}
\title{RMD_stable}
\usage{
RMD_stable(
  Y,
  D,
  X,
  p0,
  D_LB = 0,
  D_add = 0.2,
  max_iter = 10,
  b = NULL,
  c = 0.5,
  alpha = 0.1,
  tol = 1e-06
)
}
\arguments{
\item{Y}{outcome variable / dependent variable}

\item{D}{treatment variable (binary)}

\item{X}{exogenous variables}

\item{p0}{initial value of p. By default, a rule of thumb is applied.}

\item{D_LB}{lower bound on D (default 0)}

\item{D_add}{value added to D (default 0.2)}

\item{max_iter}{maximum iterations of Lasso (default 10)}

\item{c}{parameter to tune lambda (default 0.5)}

\item{alpha}{parameter to tune lambda (default 0.1)}

\item{tol}{minimum improvement to continue looping (default 1e-6)}
}
\description{
Estimation of parameters.
}
