% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlassoAutoDML_helpers.R
\name{RMD_stable}
\alias{RMD_stable}
\title{RMD_stable}
\usage{
RMD_stable(
  Y,
  D,
  X,
  p0,
  D_LB,
  D_add,
  max_iter,
  b,
  is_alpha,
  is_lasso,
  c = 0.5,
  alpha = 0.1,
  tol = 1e-06
)
}
\arguments{
\item{Y}{}

\item{D}{}

\item{X}{}

\item{p0}{}

\item{D_LB}{}

\item{D_add}{}

\item{max_iter}{}

\item{b}{}

\item{is_alpha}{}

\item{is_lasso}{}

\item{tol}{}
}
\description{
TODO: inset description on what this function does and document all
parameters
}
\details{
TODO: set defaults for arguments
}
